+++ ci/run_before_install.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-set -ev
-
-docker pull $IMAGE