+++ CHANGELOG.md
@@ -17,7 +17,7 @@ $$/   $$/ $$$$$$/ $$/       $$/           $$/     $$/ $$/ $$$$$$$/
 
 Version 3.1.0-dev
 ---------------------------------------- 
-- Added broadcast method to collectives, as well as expanding coverage to MultiVector collectives
+- Improve support for parallel sampling
 
 Version 3.0.0, released on Feb 2, 2020
 ----------------------------------------