+++ CHANGELOG.md
@@ -15,6 +15,10 @@ $$/   $$/ $$$$$$/ $$/       $$/           $$/     $$/ $$/ $$$$$$$/
 
                           https://hippylib.github.io
 
+Development version
+-------------------
+- Introduce utilities to interpolate from cartesian data (expressed as `numpy arrays`) on a `dolfin` mesh. 
+
 Version 3.0.0, released on Feb 2, 2020
 ---------------------------------------- 
 - Support for `FEniCS 2019.1.0`