+++ CHANGELOG.md
@@ -17,6 +17,7 @@ $$/   $$/ $$$$$$/ $$/       $$/           $$/     $$/ $$/ $$$$$$$/
 
 Development version
 -------------------
+- Add support for non-Gaussian noise models
 - Introduce utilities to interpolate cartesian data (expressed as `numpy arrays`) on a `dolfin` mesh. 
 
 Version 3.0.0, released on Feb 2, 2020